Modifications made to CipherSaber2
