+++ src/core/config/Categories.json
@@ -96,7 +96,8 @@
             "A1Z26 Cipher Encode",
             "A1Z26 Cipher Decode",
             "Atbash Cipher",
-            "CipherSaber2",
+            "CipherSaber2 Encrypt",
+            "CipherSaber2 Decrypt",
             "Substitute",
             "Derive PBKDF2 key",
             "Derive EVP key",
